Fix TypeError loading model: use torch_dtype, not dtype, kwarg
AutoModelForCausalLM.from_pretrained(model_name, dtype=dtype) leaks
the dtype kwarg through to AutoConfig.from_pretrained on this
transformers version, where it lands in the raw config dict. The
config's __repr__ (triggered by an internal logger.info call) then
tries to JSON-serialize it and crashes since a live torch.dtype
object isn't JSON-serializable. torch_dtype is intercepted directly
by from_pretrained's own signature, so it never reaches that path.

+++ src/dravidian_lm/evaluation/run_eval.py
@@ -157,7 +157,7 @@ def load_model_and_tokenizer(model_name: str, device: str):
     tokenizer = load_tokenizer_from_hub(model_name)
 
     dtype = torch.float16 if device.startswith("cuda") else torch.float32
-    model = AutoModelForCausalLM.from_pretrained(model_name, dtype=dtype)
+    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=dtype)
     model.to(device)
     model.eval()
